jni: refine realtime video recognition
