+++ core/ggml/CMakeLists.txt
@@ -242,6 +242,7 @@ set(SOURCE_FILES
         ${KANTV_GGMLJNI_SRC_DIR}/llm-inference.cpp
         ${KANTV_GGMLJNI_SRC_DIR}/sd-inference.cpp
         ${KANTV_GGMLJNI_SRC_DIR}/ndkcamera.cpp
+        ${KANTV_GGMLJNI_SRC_DIR}/myndkcamera.cpp
         ${KANTV_GGMLJNI_SRC_DIR}/bmp.c
         ${KANTV_GGMLJNI_SRC_DIR}/realtime-video-recognition.cpp
 )
