+++ core/ggml/jni/realtime-video-recognition.cpp
@@ -69,12 +69,7 @@ extern "C" {
 
 #include "kantv-asr.h"
 #include "ggml-jni.h"
-
-#include "whisper.h"
-
 #include "llama.h"
-
-#include "llamacpp/ggml/include/ggml-hexagon.h"
 
 #include <android/asset_manager_jni.h>
 #include <android/native_window_jni.h>
@@ -107,88 +102,10 @@ extern "C" {
 #include "speculative.h"
 #include "mtmd.h"
 
+#include "myndkcamera.h"
 
-class MyNdkCamera;
 
-static ncnn::Mutex      g_ncnn_lock;
 static MyNdkCamera *    g_camera        = nullptr;
-static long long        g_bmp_idx       = 0;
-static char             g_bmp_filename[JNI_TMP_LEN];
-
-class MyNdkCamera : public NdkCameraWindow {
-public:
-    virtual void on_image_render(cv::Mat &rgb) const;
-};
-
-static int draw_unsupported(cv::Mat &rgb) {
-    const char text[] = "unsupported";
-
-    int baseLine = 0;
-    cv::Size label_size = cv::getTextSize(text, cv::FONT_HERSHEY_SIMPLEX, 1.0, 1, &baseLine);
-
-    int y = (rgb.rows - label_size.height) / 2;
-    int x = (rgb.cols - label_size.width) / 2;
-
-    cv::rectangle(rgb, cv::Rect(cv::Point(x, y),
-                                cv::Size(label_size.width, label_size.height + baseLine)),
-                  cv::Scalar(255, 255, 255), -1);
-
-    cv::putText(rgb, text, cv::Point(x, y + label_size.height),
-                cv::FONT_HERSHEY_SIMPLEX, 1.0, cv::Scalar(0, 0, 0));
-
-    return 0;
-}
-
-static int draw_fps(cv::Mat & rgb) {
-    // resolve moving average
-    float avg_fps = 0.f;
-    {
-        static double t0 = 0.f;
-        static float fps_history[10] = {0.f};
-
-        double t1 = ncnn::get_current_time();
-        if (t0 == 0.f) {
-            t0 = t1;
-            return 0;
-        }
-
-        float fps = 1000.f / (t1 - t0);
-        t0 = t1;
-
-        for (int i = 9; i >= 1; i--) {
-            fps_history[i] = fps_history[i - 1];
-        }
-        fps_history[0] = fps;
-
-        if (fps_history[9] == 0.f) {
-            return 0;
-        }
-
-        for (int i = 0; i < 10; i++) {
-            avg_fps += fps_history[i];
-        }
-        avg_fps /= 10.f;
-    }
-
-    char text[32];
-    snprintf(text, 32,"FPS=%.2f", avg_fps);
-
-    int baseLine = 0;
-    cv::Size label_size = cv::getTextSize(text, cv::FONT_HERSHEY_SIMPLEX, 0.5, 1, &baseLine);
-
-    int y = 0;
-    int x = rgb.cols - label_size.width;
-
-    cv::rectangle(rgb, cv::Rect(cv::Point(x, y),
-                                cv::Size(label_size.width, label_size.height + baseLine)),
-                  cv::Scalar(255, 255, 255), -1);
-
-    cv::putText(rgb, text, cv::Point(x, y + label_size.height),
-                cv::FONT_HERSHEY_SIMPLEX, 0.5, cv::Scalar(0, 0, 0));
-
-    return 0;
-}
-
 
 //ref:https://github.com/ggml-org/llama.cpp/blob/master/tools/server/utils.hpp#L1300-L1309
 // Computes FNV-1a hash of the data
@@ -438,66 +355,14 @@ failure:
     return;
 }
 
-void MyNdkCamera::on_image_render(cv::Mat & rgb) const {
-    g_bmp_idx++;
-    if (0 == inference_is_running_state()) {
-        draw_fps(rgb);
-        return;
-    }
-    //TODO: stability issue between UI layer and native layer
-    if (0 == (g_bmp_idx % 100)) { // 100 / 30 ~= 3 seconds
-#if 0
-        //snprintf(g_bmp_filename, JNI_TMP_LEN, "/sdcard/bmp-%04d.bmp", g_bmp_idx);
-        snprintf(g_bmp_filename, JNI_TMP_LEN, "/sdcard/bmp-tmp.bmp");
-        LOGGD("write bmp %s, width %d, height %d\n", g_bmp_filename, rgb.cols, rgb.rows);
-
-        //this is a lazy/dirty method to implement "realtime"(not realtime at the moment) video recognition via MTMD feature in llama.cpp
-        //using MTMD API directly is a better approach, can be seen in: https://github.com/ggml-org/llama.cpp/blob/master/tools/server/server.cpp#L4121
-        write_bmp(g_bmp_filename, rgb.cols, rgb.rows, 24, rgb.data);
-        inference_init_running_state();
-        llava_inference("/sdcard/SmolVLM2-256M-Video-Instruct-f16.gguf",
-                        "/sdcard/mmproj-SmolVLM2-256M-Video-Instruct-f16.gguf",
-                        g_bmp_filename, "what do you see in this image?",
-                        GGML_BENCHMARK_LLM, 8, HEXAGON_BACKEND_GGML, HWACCEL_CDSP);
-        inference_reset_running_state();
-#else
-        //better approach: using MTMD API directly to avoid write bmp data to storage
-        mtmd_inference(rgb);
-#endif
-    }
-
-    draw_fps(rgb);
-}
-
-
-extern "C" {
-
-JNIEXPORT jint JNI_OnLoad(JavaVM * vm, void * reserved) {
-    LOGGD("JNI_OnLoad");
-
-    ncnn::create_gpu_instance();
-    if (nullptr == g_camera)
-        g_camera = new MyNdkCamera;
-
-    return JNI_VERSION_1_6;
-}
-
-JNIEXPORT void JNI_OnUnload(JavaVM * vm, void * reserved) {
-    LOGGD("JNI_OnUnload");
-    ncnn::destroy_gpu_instance();
-    if (nullptr != g_camera) {
-        delete g_camera;
-        g_camera = nullptr;
-    }
-}
-
-JNIEXPORT jboolean JNICALL
-Java_kantvai_ai_ggmljava_openCamera(JNIEnv * env, jclass clazz, jint facing) {
+bool jni_open_camera(int facing) {
     if (facing < 0 || facing > 1)
-        return JNI_FALSE;
+        return false;
 
     LOGGD("openCamera %d", facing);
+
     if (nullptr == g_camera) {
+        ncnn::create_gpu_instance();
         g_camera = new MyNdkCamera;
         g_camera->open((int) facing);
         inference_reset_running_state();
@@ -505,13 +370,13 @@ Java_kantvai_ai_ggmljava_openCamera(JNIEnv * env, jclass clazz, jint facing) {
         LOGGD("camera already opened");
     }
 
-    return JNI_TRUE;
+    return true;
 }
 
-JNIEXPORT void JNICALL
-Java_kantvai_ai_ggmljava_closeCamera(JNIEnv * env, jclass clazz) {
+void jni_close_camera() {
     LOGGD("closeCamera");
     if (nullptr != g_camera) {
+        ncnn::destroy_gpu_instance();
         inference_reset_running_state();
         g_camera->close();
         //FIXME:comment following line to avoid deadlock but brings memory leak
@@ -522,10 +387,7 @@ Java_kantvai_ai_ggmljava_closeCamera(JNIEnv * env, jclass clazz) {
     }
 }
 
-JNIEXPORT void JNICALL
-Java_kantvai_ai_ggmljava_setOutputWindow(JNIEnv * env, jclass clazz, jobject surface) {
-    ANativeWindow *win = ANativeWindow_fromSurface(env, surface);
-
+void jni_set_outputwindow(ANativeWindow  * win) {
     LOGGD("setOutputWindow %p", win);
     if (nullptr != g_camera) {
         LOGGD("setOutputWindow %p", win);
@@ -533,4 +395,28 @@ Java_kantvai_ai_ggmljava_setOutputWindow(JNIEnv * env, jclass clazz, jobject sur
     }
 }
 
-} //end extern "C" {
+void multimodal_inference(cv::Mat & rgb) {
+    static long long        g_bmp_idx       = 0;
+    g_bmp_idx++;
+    if (0 == inference_is_running_state()) {
+        return;
+    }
+    if (0 == (g_bmp_idx % 100)) { // 100 / 30 ~= 3 seconds
+#if 0
+        static char             g_bmp_filename[JNI_TMP_LEN];
+        //snprintf(g_bmp_filename, JNI_TMP_LEN, "/sdcard/bmp-%04d.bmp", g_bmp_idx);
+        snprintf(g_bmp_filename, JNI_TMP_LEN, "/sdcard/bmp-tmp.bmp");
+        LOGGD("write bmp %s, width %d, height %d\n", g_bmp_filename, rgb.cols, rgb.rows);
+        //this is a lazy/dirty method to implement "realtime"(not realtime at the moment) video recognition via MTMD feature in llama.cpp
+        //using MTMD API directly is a better approach, can be seen in: https://github.com/ggml-org/llama.cpp/blob/master/tools/server/server.cpp#L4121
+        write_bmp(g_bmp_filename, rgb.cols, rgb.rows, 24, rgb.data);
+        llava_inference("/sdcard/SmolVLM2-256M-Video-Instruct-f16.gguf",
+                        "/sdcard/mmproj-SmolVLM2-256M-Video-Instruct-f16.gguf",
+                        g_bmp_filename, "what do you see in this image?",
+                        GGML_BENCHMARK_LLM, 8, HEXAGON_BACKEND_GGML, HWACCEL_CDSP);
+#else
+        //better approach: using MTMD API directly to avoid write bmp data to storage
+        mtmd_inference(rgb);
+#endif
+    }
+}
